Extract token deposit options to own hook
